backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Async Tasks
celery==5.3.4
kombu==5.3.4

# SMS & Email
twilio==8.10.0
africastalking==1.2.6
sendgrid==6.11.0

# Media & Storage
boto3==1.29.7
cloudinary==1.36.0
Pillow==10.1.0
python-magic==0.4.27

# WebSockets
websockets==12.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.26

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

